BambiCTF_2021/ringring/App/requirements.txt
Babel==2.8.0
click==7.1.2
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Table==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pytz==2020.1
urllib3==1.25.8
requests==2.23.0
Werkzeug==1.0.1
gunicorn[eventlet]